requirements.txt
alembic==1.15.1
appdirs==1.4.4
axios==0.4.0
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cmake==3.31.6
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
Deprecated==1.2.18
dnspython==2.7.0
eventlet==0.39.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
limits==5.4.0
llvmlite==0.44.0
lxml==5.3.1
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
ndg-httpsclient==0.5.1
networkx==3.5
ninja==1.11.1.3
numba==0.61.2
numpy==2.1.3
openai-whisper==20240930
ordered-set==4.1.0
packaging==24.2
pillow==10.4.0
proglog==0.1.10
psycopg2==2.9.10
pusher==3.3.3
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.22
pyee==11.1.1
Pygments==2.19.1
PyNaCl==1.5.0
pyOpenSSL==25.0.0
pyppeteer==2.0.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
redis==5.2.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
setuptools==75.8.2
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.38
srt==3.5.3
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==1.26.20
vosk==0.3.45
websockets==10.4
Werkzeug==3.1.3
wrapt==1.17.2
wsproto==1.2.0
zipp==3.21.0
